% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_ML_FIlter.R
\name{CatBoost_feature_selection}
\alias{CatBoost_feature_selection}
\title{Feature selection using CatBoost algorithm}
\usage{
CatBoost_feature_selection(data, Labels, n_genes_to_keep)
}
\arguments{
\item{data}{A matrix or data frame of gene expression data (rows are genes, columns are samples).}

\item{Labels}{A vector of sample labels (factors or integers) corresponding to the columns in the data.}

\item{n_genes_to_keep}{The number of top genes to keep based on feature importance.}
}
\value{
A list containing:
\itemize{
\item FilteredData: A data frame or matrix containing only the top n_genes_to_keep genes.
\item Important_Features: A vector of the top n_genes_to_keep gene names.
\item FeatureImportance: A data frame containing the feature importance scores for all genes.
}
}
\description{
This function performs feature selection on single-cell RNA-seq data using the CatBoost algorithm.
It trains a CatBoost model, calculates feature importance, and returns the top n_genes_to_keep genes.
}
