% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFilterMethods.R
\name{CorrelationFun}
\alias{CorrelationFun}
\title{Correlation-Based Feature Selection}
\usage{
CorrelationFun(Data, Labels, n_genes_to_keep)
}
\arguments{
\item{Data}{A data frame or matrix containing gene expression values. Rows represent genes,
and columns represent samples.}

\item{Labels}{A vector containing the labels of the samples. Must have the same length as
the number of columns in the Data.}

\item{n_genes_to_keep}{The number of top genes to keep based on the absolute correlation coefficients.}
}
\value{
A list containing the following elements:
\itemize{
\item{FilterData}{A data frame or matrix containing the selected genes.}
\item{Important_Features}{A vector of the names of the selected genes.}
\item{Correlation_Analysis}{A data frame containing the absolute correlation coefficients and gene names.}
}
}
\description{
This function applies Pearson's correlation coefficient to select the top n genes
with the highest absolute correlation coefficients.
}
