% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_VariableGenesMethods.R
\name{DUBStepRfun}
\alias{DUBStepRfun}
\title{DUBStepRfun}
\usage{
DUBStepRfun(data, K, Num.pcs)
}
\arguments{
\item{data}{A data frame or matrix containing gene expression values.}

\item{K}{An integer indicating the number of nearest neighbors to use in determining cells that may be drop-outs.}

\item{Num.pcs}{An integer indicating the number of principal components to use for clustering.}
}
\value{
A list of the filtered gene expression data and a list of DUBStepR output results.
}
\description{
This function filters cells using DUBStepR, which is an R package that enables identification of "drop-outs" in single-cell transcriptome analysis.
}
