% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataset.R
\docType{data}
\name{ExampleDataset}
\alias{ExampleDataset}
\title{Example Single-Cell RNA Expression Dataset}
\format{
A data.frame with rows representing genes and columns
representing cells:
\describe{
\item{Cells}{Numeric vector of expression values for cells}
...
}
}
\usage{
data(ExampleDataset)
}
\description{
A data.frame object that represents gene expression data from
single-cell RNA sequencing. The rows in the data.frame correspond to
individual genes, while the columns correspond to individual cells.
}
\keyword{datasets}
