% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_FS_Methods.R
\name{Filter_FS_Methods}
\alias{Filter_FS_Methods}
\title{Filter Feature Selection Methods}
\usage{
Filter_FS_Methods(
  data,
  Labels,
  n_genes_to_keep,
  method,
  LogTransformation = TRUE,
  HighVariableFIlter = TRUE,
  n_features = 2000
)
}
\arguments{
\item{data}{Data frame/matrix of gene expression values. Rows are genes and
columns are samples.}

\item{Labels}{Vector of sample labels. Length must equal the number of data columns.}

\item{n_genes_to_keep}{Integer. Number of top genes to retain after feature
selection.}

\item{method}{Character. Name of the feature selection method to apply. Valid
methods: "WaldTest", "BetaPoissonGLM", "WilcoxonTest", "LikelihoodRatioTest",
"KruskalWallis", "mRMRe", "QuasiLikelihoodFTest", "KendallCorrelation",
"SpearmanCorrelation", "InformationGain".}

\item{LogTransformation}{Logical. If TRUE, apply log transformation on the data
using Seurat. Default is TRUE.}

\item{HighVariableFIlter}{Logical. If TRUE, apply high variability gene filtering.
Default is TRUE.}

\item{n_features}{Integer. Number of genes with high variability to retain.}
}
\value{
Data frame/matrix of top n_genes_to_keep genes selected by the specified
method.
}
\description{
This function applies selected feature selection methods to single-cell RNA
sequencing data. It takes as input a gene expression matrix, sample labels,
the number of top genes to retain, and the name of the feature selection method.
Optional parameters include log transformation and high variability gene filtering.
The function returns a matrix of the top genes selected by the specified method.
Available feature selection methods include statistical tests
(Wald Test, Beta Poisson GLM, Wilcoxon Test, Likelihood Ratio Test, Kruskal Wallis),
correlation measures (Kendall, Spearman), information gain, and mRMRe.
This function is designed to facilitate the comparison of different feature
selection methods and to assist in the identification of the most informative genes
for further analysis.
}
\examples{
# Example data and labels

# Apply the Wald Test method and keep the top 10 genes
result <- Filter_FS_Methods(ExampleDataset[1:200, ], Labels,
    n_genes_to_keep = 50, method = "WaldTest",
    LogTransformation = TRUE, HighVariableFIlter = TRUE, n_features = 100)
}
