% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessFilter.R
\name{filter_genes}
\alias{filter_genes}
\title{Apply a statistical filter to gene expression data}
\usage{
filter_genes(data, filter_method, n_features, unique_cut)
}
\arguments{
\item{data}{A matrix or data.frame object representing gene expression data. Rows represent genes, and columns represent samples.}

\item{filter_method}{A character string indicating the method to use for filtering. Can be one of "High_Variable_Genes" (to use the Seurat FindVariableFeatures method) or "Remove_Low_Variance" (to use the zero variance filter).}

\item{n_features}{An integer indicating the number of genes to keep after filtering using the Seurat FindVariableFeatures method. Default is 500.}

\item{unique_cut}{A numeric cutoff threshold for the zero variance filter. Features with a variance below this threshold will be filtered out.}
}
\value{
A filtered dataset with only the genes that pass the statistical filter.
\item{filtered_data}{A matrix or data.frame object representing the filtered gene expression data.}
\item{important_genes}{A character vector of gene names that pass the statistical filter.}
\item{statistical_analysis}{A data.frame object containing the p-values or FDRs for all genes, as calculated by the selected statistical method.}
}
\description{
This function performs a statistical filtering method on gene expression data. It filters the genes based on their statistical significance calculated by various methods such as Seurat, DESeq2, MAST, or BPSC.
}
\examples{
\donttest{
data(ExampleDataset)
data_filtered <- filter_genes(ExampleDataset,filter_method="High_Variable_Genes",n_features= 100)
}
}
