% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterData.R
\docType{data}
\name{FilterData}
\alias{FilterData}
\title{Results of filtered processing}
\format{
A matrix or data.frame object representing the gene expression data
after it has been filtered or processed in some way.
\describe{
\item{Cells}{Numeric vector of expression values for cells}
...
}
}
\usage{
FilterData
}
\description{
A list object that represents the gene expression data after it
has been filtered or processed in some way.
}
\examples{
data(FilterData)
}
\keyword{datasets}
