% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFilterMethods.R
\name{ig_fun}
\alias{ig_fun}
\title{Information Gain-Based Feature Selection for Single-Cell RNA-seq Data}
\usage{
ig_fun(Data, Labels, n_genes_to_keep)
}
\arguments{
\item{Data}{A data frame or matrix containing the single-cell RNA-seq expression data.}

\item{Labels}{A vector containing the labels for each sample in the input data.}

\item{n_genes_to_keep}{An integer specifying the number of top important features (genes) to keep.}
}
\value{
A list containing the following elements:
\itemize{
\item{FilterData}{A data frame or matrix containing the filtered input data, keeping only the top important features.}
\item{Important_Features}{A character vector of the top important feature (gene) names.}
\item{Information_Gain_Analysis}{A data frame containing the Information Gain scores for each feature (gene).}
}
}
\description{
This function selects the top important features (genes) based on their
Information Gain scores from single-cell RNA-seq data.
}
