% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFilterMethods.R
\name{KendallFun}
\alias{KendallFun}
\title{Kendall Rank Correlation Feature Selection}
\usage{
KendallFun(data, Labels, n_genes_to_keep)
}
\arguments{
\item{data}{A data frame or matrix containing gene expression values. Rows represent genes,
and columns represent samples.}

\item{Labels}{A vector containing the labels of the samples. Must have the same length as
the number of columns in the data.}

\item{n_genes_to_keep}{The number of top genes to keep based on the FDR-adjusted p-values.}
}
\value{
A list containing the following elements:
\itemize{
\item{FilterData}{A data frame or matrix containing the selected genes.}
\item{Important_Features}{A vector of the names of the selected genes.}
\item{Statistical_Analysis}{A data frame containing the p-values, FDR-adjusted p-values, and gene names.}
}
}
\description{
This function applies Kendall's rank correlation coefficient to select the top n genes
with the lowest FDR-adjusted p-values.
}
