% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\docType{data}
\name{Labels}
\alias{Labels}
\title{Cell Labels}
\format{
An object of class \code{factor} of length 384.
}
\usage{
Labels
}
\value{
A character or factor vector of sample labels, with one label per
column of \code{data}.
}
\description{
A character or factor vector specifying the label for each cell
in the input data matrix.
}
\examples{
data(Labels)

}
\keyword{cell}
\keyword{labels}
\keyword{labels,}
\keyword{sample}
