% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_ML_FIlter.R
\name{ML_filter}
\alias{ML_filter}
\title{Machine Learning-based Feature Selection}
\usage{
ML_filter(data, Labels, MLmethod, n_genes_to_keep)
}
\arguments{
\item{data}{A data frame or matrix containing gene expression values. Rows represent genes and columns represent samples.}

\item{Labels}{A vector containing the labels of the samples. Must be the same length as the number of columns in the data.}

\item{MLmethod}{Character, the name of the machine learning method to be applied. Supported methods include:
"knn", "svmLinear", "lda", "rf", "rpart", "C5.0", "gbm", and "xgbTree".}

\item{n_genes_to_keep}{Integer, the number of top genes to keep after applying the selected machine learning method.}
}
\value{
A list containing the following elements:
\itemize{
\item FilterData: A data frame or matrix containing the top n_genes_to_keep genes selected by the specified machine learning method.
\item Important_Features: A vector of the names of the top n_genes_to_keep genes.
\item ML_Analysis: A data frame containing the importance scores of all genes considered by the machine learning method.
}
}
\description{
This function applies machine learning methods to select top genes based on their importance in classification tasks.
}
