% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_FS_Methods.R
\name{ML_FS_Methods}
\alias{ML_FS_Methods}
\title{ML-Based Feature Selection Methods}
\usage{
ML_FS_Methods(
  data,
  Labels,
  n_genes_to_keep,
  method,
  LogTransformation = TRUE,
  HighVariableFIlter = TRUE,
  n_features = 2000
)
}
\arguments{
\item{data}{Matrix or data frame of input data (features and samples).}

\item{Labels}{Vector of true class labels of the samples.}

\item{n_genes_to_keep}{Integer. Number of top-ranked genes to retain.}

\item{method}{Character. ML method for feature selection. Supported methods:
"PAM", "rangerBagging", "PLS", "RF", "CART", "C5.0", "GBM", "xgbTree", "Catboost",
"glmnet".}

\item{LogTransformation}{Logical. If TRUE, apply log transformation using Seurat.
Default is TRUE.}

\item{HighVariableFIlter}{Logical. If TRUE, apply high variability gene filtering.
Default is TRUE.}

\item{n_features}{Integer. Number of genes with high variability to retain.}
}
\value{
List containing the filtered data, key features, and other information
relevant to the selected method.
}
\description{
This function applies machine learning (ML) methods for feature selection in single-cell
RNA-seq data. It takes as input a gene expression matrix, sample labels,
the number of top genes to retain, and the name of the ML method.
Optional parameters include log transformation and high variability gene filtering.
The function returns a list containing the filtered data, key features, and other
information relevant to the selected method.
Available ML methods include PAM, rangerBagging, PLS, RF, CART, C5.0, GBM, xgbTree,
Catboost, and glmnet. These methods can help identify the most informative genes
for downstream analysis.
This function is designed to facilitate the comparison of different ML methods
and to assist in the identification of the most informative genes for further analysis.
}
\examples{
# Example usage:
# Apply the rangerBagging method and keep the top 100 genes
result <- ML_FS_Methods(ExampleDataset[1:200, ], Labels,
    n_genes_to_keep = 50, method = "rangerBagging",
    LogTransformation = TRUE, HighVariableFIlter = TRUE, n_features = 100)
}
