% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFilterMethods.R
\name{mRMReFun}
\alias{mRMReFun}
\title{mRMReFun: Minimum Redundancy Maximum Relevance Feature Selection}
\usage{
mRMReFun(data, Labels, n_genes_to_keep)
}
\arguments{
\item{data}{A data frame or matrix containing the expression values with rows as features/genes and columns as samples.}

\item{Labels}{A character or factor vector containing the labels for each sample.}

\item{n_genes_to_keep}{An integer specifying the number of top features to select.}
}
\value{
A list containing the following elements:
\itemize{
\item{FilterData}{A filtered data frame containing only the selected features/genes.}
\item{Important_Features}{A character vector of the selected feature/genes names.}
}
}
\description{
This function performs Minimum Redundancy Maximum Relevance (mRMR) feature selection on a given dataset.
It selects the top \code{n_genes_to_keep} features based on the mRMR criterion.
}
