% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_Tree_Based_ML.R
\name{Non_Tree_Based_ML}
\alias{Non_Tree_Based_ML}
\title{Non-Tree-Based Machine Learning for Feature Selection in scRNA-seq Data}
\usage{
Non_Tree_Based_ML(data, Labels, MLmethod, n_genes_to_keep)
}
\arguments{
\item{data}{A numeric matrix of gene expression values with rows representing
genes and columns representing cells.}

\item{Labels}{A character or factor vector specifying the label for each cell
in \code{data}.}

\item{MLmethod}{A character string specifying the non-tree machine learning
method to be used. The valid options are "knn", "svmRadial", "glmnet", "lda".}

\item{n_genes_to_keep}{An integer value that sets the number of top-ranked
features to keep. Only the top \code{n_genes_to_keep} features with the
highest importance scores will be retained.}
}
\value{
A list with three elements: \code{FilterData}, \code{Important_Features}
and \code{ML_Analysis}.
\code{FilterData} is a numeric matrix of gene expression values with rows
representing genes and columns representing cells, after the feature selection
process.
\code{Important_Features} is a vector of important genes selected by the
non-tree machine learning-based feature selection method.
\code{ML_Analysis} is a data frame containing the importance scores for all
features used in the non-tree machine learning analysis.
}
\description{
This function applies non-tree-based machine learning-based feature selection
to single-cell RNA sequencing (scRNA-seq) data. The function takes in a data
matrix, a vector of cell labels, a machine learning method, and additional
parameters to configure the machine learning method. The resulting filtered
data, important features, and the ML analysis are returned as a list.
}
\examples{

data(ExampleDataset)
Non_Tree_Based_ML(head(ExampleDataset, 10),Labels<-Labels,"knn",
n_genes_to_keep <- 100)

}
