% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessFilter.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize a matrix of gene expression data using different methods}
\usage{
normalize_data(ExpressionData, normal_method)
}
\arguments{
\item{ExpressionData}{A Table of gene expression data to be normalized. Rows
represent genes, and columns represent samples.}

\item{normal_method}{A character string specifying the normalization method.
Possible values are "TMM", "RLE", "CPM", "TPM", and "LogNormalize".}
}
\value{
A matrix of normalized gene expression data.
}
\description{
This function normalizes a matrix of gene expression data using different
methods, including TMM, RLE, CPM, TPM, and LogNormalize.
}
\examples{

data(ExampleDataset)
Norm_data <- normalize_data(ExpressionData <-
 head(ExampleDataset, 100),normal_method <- "LogNormalize")

}
