% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_StatisticalMethods.R
\name{PvalueCalc}
\alias{PvalueCalc}
\title{PvalueCalc function}
\usage{
PvalueCalc(
  data,
  Pvaluemethod,
  Labels,
  Test = "LRT",
  dmethod = "bimod",
  DESeq2.test = "LRT",
  method_MAST = "bayesglm",
  MAST.parallel = TRUE
)
}
\arguments{
\item{data}{a SummarizedExperiment object containing gene expression data}

\item{Pvaluemethod}{a character string specifying the method to be used for calculating p-values and FDRs. Possible values are "Seurat_method", "BPSC_metchod", "DESeq2_method", and "MAST_method".}

\item{Labels}{a vector of group labels for the samples in the data object}

\item{Test}{a character string specifying the test to be used in the DESeq2 method. Default is "LRT".}

\item{dmethod}{a character string specifying the method to be used in the Seurat method. Default is "bimod".}

\item{DESeq2.test}{a character string specifying the test to be used in the DESeq2 method. Default is "LRT".}

\item{method_MAST}{a character string specifying the method to be used in the MAST method. Default is "bayesglm".}

\item{MAST.parallel}{a logical value indicating whether parallel computing should be used in the MAST method. Default is TRUE.}
}
\description{
This function calculates p-values and FDRs for differential expression analysis using different methods.
}
