% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFilterMethods.R
\name{QLFTfun}
\alias{QLFTfun}
\title{QLFTfun
Filter genes using Quasi-Likelihood F-test (QLFT) from the edgeR package.}
\usage{
QLFTfun(data, Labels, n_genes_to_keep)
}
\arguments{
\item{data}{A SummarizedExperiment object containing gene expression data.}

\item{Labels}{A vector of character or factor labels for the samples.}

\item{n_genes_to_keep}{The number of top genes to keep based on the QLFT ranking.}
}
\value{
A list containing:
\itemize{
\item{FilterData}{A filtered data matrix with the top n_genes_to_keep genes.}
\item{Important_Features}{A vector of gene names for the top n_genes_to_keep genes.}
\item{Statistical_Analysis}{A data frame containing the genes, P_values, and FDR for the QLFT analysis.}
}
}
\description{
QLFTfun
Filter genes using Quasi-Likelihood F-test (QLFT) from the edgeR package.
}
