% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_VariableGenesMethods.R
\name{scDDfun}
\alias{scDDfun}
\title{Differential expression analysis using the scDD method}
\usage{
scDDfun(
  data,
  prior_param = list(alpha = 0.01, mu0 = 0, s0 = 0.01, a0 = 0.01, b0 = 0.01),
  labels,
  nGenes = 150
)
}
\arguments{
\item{data}{A numeric matrix of raw count data.}

\item{prior_param}{A list of prior parameters used for Bayesian inference. Default values are alpha=0.01, mu0=0, s0=0.01, a0=0.01, b0=0.01.}

\item{labels}{A vector of condition labels corresponding to the columns of the count matrix.}

\item{nGenes}{The number of top-ranked genes to return. Default is 150.}
}
\value{
A list with the following components:
\item{FilterData}{A filtered version of the input data, retaining only genes that were found to be differentially expressed.}
\item{Important_Features}{A vector of the names of the genes that were found to be differentially expressed.}
\item{Results}{A data frame with the results of the differential expression analysis.}
}
\description{
This function performs differential expression analysis on single-cell RNA sequencing data using the scDD method. It takes as input a count matrix, a list of prior parameters, and a vector of condition labels. The function returns a list of filtered data, important features, and results of the analysis.
}
