% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_VariableGenesMethods.R
\name{scmapfun}
\alias{scmapfun}
\title{Perform feature selection and identify important features using scmap}
\usage{
scmapfun(data, Labels, nGenes)
}
\arguments{
\item{data}{A matrix or data frame containing the gene expression data for the single cells.
Rows represent genes, and columns represent cells.}

\item{Labels}{A vector containing the cell labels.
Rows represent cells, and columns represent metadata associated with the cells (e.g.,
sample ID, batch, cell type, etc.).}

\item{nGenes}{An integer specifying the number of genes to select as important features.}
}
\value{
A list containing the following items:
\describe{
\item{FilterData}{A matrix containing the filtered gene expression data. Rows represent genes,
and columns represent cells. Only the nGenes genes selected as important features are included.}
\item{Important_Features}{A character vector containing the names of the nGenes genes selected
as important features.}
\item{Results}{A data frame containing the output of the scmap analysis. Rows represent genes,
and columns represent various statistics (e.g., scores, p-values, etc.) calculated by scmap.}
}
}
\description{
This function performs feature selection and identifies important features based
on the output of the scmap package. The scmap package is used to find the most
informative genes for separating two groups of cells, and these genes are selected
as important features. The function returns a list containing the filtered data,
the list of important features, and the results of the scmap analysis.
}
