% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_VariableGenesMethods.R
\name{SCMarkerfun}
\alias{SCMarkerfun}
\title{Identify important genes in single-cell RNA-seq data using SCMarker}
\usage{
SCMarkerfun(data, GeneSK, CellSK)
}
\arguments{
\item{data}{A matrix with rows representing genes and columns representing cells.}

\item{GeneSK}{An integer indicating the number of genes to use in the clustering.}

\item{CellSK}{An integer indicating the number of cells to use in the clustering.}
}
\value{
A list containing the SCMarker object and a vector of important genes.
}
\description{
This function uses the SCMarker package to identify important genes in single-cell RNA-seq data.
}
