% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_VariableGenesMethods.R
\name{scPNMFfun}
\alias{scPNMFfun}
\title{Perform scPNMF filtering on single-cell data}
\usage{
scPNMFfun(data, DM, gN)
}
\arguments{
\item{data}{A numeric matrix of gene expression values}

\item{DM}{The distance metric to be used in the scPNMF algorithm (either "KL" for Kullback-Leibler divergence or "JS" for Jensen-Shannon divergence)}

\item{gN}{The number of genes to be retained in the filtered dataset}
}
\value{
A list with two elements:
\describe{
\item{FilterData}{The filtered single-cell expression data}
\item{Important_Features}{The selected genes with the highest relevance scores}
}
}
\description{
Perform scPNMF filtering on single-cell data
}
