% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_VariableGenesMethods.R
\name{SelfEGenes}
\alias{SelfEGenes}
\title{Select top self-expressed genes}
\usage{
SelfEGenes(data, n)
}
\arguments{
\item{data}{A matrix or data frame where rows are samples and columns are genes.}

\item{n}{An integer indicating the number of top self-expressed genes to select.}
}
\value{
A list with two elements:
\describe{
\item{FilterData}{A data frame containing the top self-expressed genes.}
\item{Important_Features}{A character vector containing the names of the selected genes.}
}
}
\description{
This function selects the top self-expressed genes based on the number of times a gene is its own nearest neighbor in a gene expression dataset.
}
