% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_ML_FIlter.R
\name{ShapleyValuesFun}
\alias{ShapleyValuesFun}
\title{ShapleyValues Function}
\usage{
ShapleyValuesFun(data, Labels, n_genes_to_keep)
}
\arguments{
\item{data}{The input dataset, a matrix or data frame of numeric values.}

\item{Labels}{A vector of labels for each row in the input data.}

\item{n_genes_to_keep}{The number of top features to keep.}
}
\value{
A list containing three elements:
\describe{
\item{FilterData}{The filtered dataset with only the top features.}
\item{Important_Features}{The names of the top features.}
\item{Statistical_Analysis}{A data frame with the Shapley values for all the features.}
}
}
\description{
This function calculates the Shapley values of a set of features for a given dataset.
}
