% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_VariableGenesMethods.R
\name{SIMLRFun}
\alias{SIMLRFun}
\title{Run SIMLR clustering and feature ranking on single-cell RNA-seq data}
\usage{
SIMLRFun(data, ClusterNumber = 3, nGenes = 150)
}
\arguments{
\item{data}{A matrix of gene expression values, where rows represent genes and columns represent cells.}

\item{ClusterNumber}{The number of cell clusters to identify.
that all available cores should be used.}

\item{nGenes}{The number of top-ranked genes to return. Default is 150.}
}
\value{
A list containing the filtered gene expression data, the list of top-ranked genes,
and the SIMLR feature ranking results.
}
\description{
This function runs the SIMLR algorithm on single-cell RNA-seq data to identify cell clusters
and rank the most informative genes. The function takes as input a matrix of gene expression
values, the number of clusters to identify, the ratio of cores to use for parallel processing,
and the number of top-ranked genes to return.
}
