% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable_Filter.R
\name{Variable_Filter}
\alias{Variable_Filter}
\title{Variable_Filter function for feature selection in scRNA-seq data}
\usage{
Variable_Filter(data, method, ...)
}
\arguments{
\item{data}{A numeric matrix of gene expression values with rows representing genes and columns representing cells.}

\item{method}{A character string specifying the feature selection method to be used. The valid options are "SCMarker", "SelfE", "DUBStepR", "scPNMF", or "M3Drop".}

\item{...}{Additional arguments that are required for the specific feature selection method chosen.}
}
\value{
A list with two elements: \code{FilterData} and \code{Important_Features}.
\code{FilterData} is a numeric matrix of gene expression values with rows representing genes and columns representing cells, after the feature selection process.
\code{Important_Features} is a vector of important genes selected by the feature selection method.
}
\description{
This function allows the user to apply various feature selection methods to single-cell RNA sequencing (scRNA-seq) data.
The function takes in a data matrix, a vector of cell labels, a feature selection method, and any additional arguments required by the method.
The resulting filtered data and important features are returned as a list.
}
\examples{
\donttest{
data(ExampleDataset)
result <- Variable_Filter(ExampleDataset,method = "DUBStepR",K=10,Num.pcs=10)
}
}
