% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataset.R
\docType{data}
\name{ExampleDataset}
\alias{ExampleDataset}
\title{Single-Cell RNA Expression}
\format{
An object of class \code{data.frame} with 2000 rows and 384 columns.
}
\usage{
ExampleDataset
}
\value{
A data.frame object representing gene expression data from single-cell RNA sequencing.
}
\description{
A data.frame object that represents gene expression data from single-cell RNA sequencing. The rows in the data.frame correspond to individual genes, while the columns correspond to individual cells.
}
\examples{
data(ExampleDataset)

}
\keyword{RNA}
\keyword{data,}
\keyword{data.frame}
\keyword{expression}
\keyword{gene}
\keyword{sequencing,}
\keyword{single-cell}
