% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_FS_Methods.R
\name{Filter_FS_Methods}
\alias{Filter_FS_Methods}
\title{Filter Feature Selection Methods}
\usage{
Filter_FS_Methods(
  data,
  Labels,
  n_genes_to_keep,
  method,
  LogTransformation = TRUE,
  HighVariableFIlter = TRUE,
  n_features = n_features
)
}
\arguments{
\item{data}{A data frame or matrix containing gene expression values. Rows represent genes and columns represent samples.}

\item{Labels}{A vector containing the labels of the samples. Must be the same length as the number of columns in the data.}

\item{n_genes_to_keep}{Integer, the number of top genes to keep after applying the selected feature selection method.}

\item{method}{Character, the name of the feature selection method to be applied. Valid methods include:
"WaldTest", "BetaPoissonGLM", "WilcoxonTest", "LikelihoodRatioTest", "KruskalWallis", "mRMRe",
"QuasiLikelihoodFTest", "KendallCorrelation", "SpearmanCorrelation", and "InformationGain".}

\item{LogTransformation}{Logical, if TRUE log transformation is applied on the data using the Seurat package. Default is TRUE.}

\item{HighVariableFIlter}{Logical, if TRUE high variability gene filtering is applied. Default is TRUE.}

\item{n_features}{Integer, the number of genes with high variability to keep.}
}
\value{
A data frame or matrix containing the top n_genes_to_keep genes selected by the specified method.
}
\description{
This function applies various feature selection methods for single-cell RNA sequencing data.
}
\examples{
# Example data and labels
data(ExampleDataset)
data(Labels)

# Apply the Wald Test method and keep the top 10 genes
result <- Filter_FS_Methods(ExampleDataset, Labels, n_genes_to_keep=100, method="WaldTest",
                            LogTransformation=TRUE, HighVariableFIlter=TRUE, n_features=2000)
}
