% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessFilter.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes according to the specified method}
\usage{
filter_genes(data, filter_method, n_features, unique_cut)
}
\arguments{
\item{data}{A numeric matrix of gene expression values with rows representing
genes and columns representing cells.}

\item{filter_method}{A character string specifying the filtering method to be
used. The valid options are "High_Variable_Genes" and "Remove_Low_Variance".}

\item{n_features}{An integer value that sets the number of top-ranked variable
genes to keep. Only used when filter_method is "High_Variable_Genes".}

\item{unique_cut}{A numeric value that sets the threshold for low variance
genes. Only used when filter_method is "Remove_Low_Variance".}
}
\value{
A filtered data frame with rows representing genes and columns
representing cells.
}
\description{
This function filters genes in scRNA-seq data according to the specified
method. The user can choose between two filtering methods:
"High_Variable_Genes" and "Remove_Low_Variance". If "High_Variable_Genes" is
selected, the function uses the FindVariableFeatures function of Seurat to
identify the top-ranked variable genes. If "Remove_Low_Variance" is selected,
the function removes genes with low variance using the nearZeroVar function
of caret.
}
\examples{

data(ExampleDataset)
data_filtered <- filter_genes(head(ExampleDataset, 100),
  filter_method<- "High_Variable_Genes", n_features<-80)

}
