% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterData.R
\docType{data}
\name{FilterData}
\alias{FilterData}
\title{Results of filtered processing}
\format{
An object of class \code{list} of length 3.
}
\usage{
FilterData
}
\value{
A matrix or data.frame object representing the gene expression data
after it has been filtered or processed in some way.
}
\description{
A list object that represents the gene expression data after it
has been filtered or processed in some way.
}
\examples{
data(FilterData)

}
\keyword{data}
\keyword{data,}
\keyword{expression}
\keyword{filtering,}
\keyword{gene}
\keyword{processing}
