% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\docType{data}
\name{Labels}
\alias{Labels}
\title{Cell Labels}
\format{
A character or factor vector of sample labels, with one label per
column of the input data matrix.
}
\usage{
data(Labels)
}
\description{
A character or factor vector specifying the label for each cell
in the input data matrix.
}
\keyword{datasets}
