% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_FS_Methods.R
\name{ML_FS_Methods}
\alias{ML_FS_Methods}
\title{ML Based Feature Selection Methods}
\usage{
ML_FS_Methods(
  data,
  Labels,
  n_genes_to_keep,
  method,
  LogTransformation = TRUE,
  HighVariableFIlter = TRUE,
  n_features = 2000
)
}
\arguments{
\item{data}{A matrix or data frame containing the input data (features and samples).}

\item{Labels}{A vector containing the true class labels of the samples.}

\item{n_genes_to_keep}{An integer representing the number of top-ranked genes to retain based on the selected method.}

\item{method}{A character string specifying the machine learning method to use for feature selection. Supported methods:
"PAM", "gcvEarth", "PLS", "RF", "CART", "C5.0", "GBM", "xgbTree", "Catboost", and "ShapleyValues".}

\item{LogTransformation}{Logical, if TRUE log transformation is applied on the data using the Seurat package. Default is TRUE.}

\item{HighVariableFIlter}{Logical, if TRUE high variability gene filtering is applied. Default is TRUE.}

\item{n_features}{Integer, the number of genes with high variability to keep.}
}
\value{
A list containing the filtered data, important features, and any additional information relevant to the selected method.
}
\description{
This function applies a specified machine learning method to filter and select important genes from single-cell RNA-seq data.
}
\examples{
# Example usage:
data(ExampleDataset)
data(Labels)
# Apply the knn method and keep the top 10 genes
result <- ML_FS_Methods(ExampleDataset, Labels, n_genes_to_keep=100, method="PAM",
                            LogTransformation=TRUE, HighVariableFIlter=TRUE, n_features=2000)
}
