% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_Analysis.R
\name{Ontology_Analysis}
\alias{Ontology_Analysis}
\title{Perform ontology analysis using clusterProfiler}
\usage{
Ontology_Analysis(
  Important_Features,
  ontology_type = "KEGG",
  organism = "human"
)
}
\arguments{
\item{Important_Features}{A character vector containing the names of the
important features (e.g., genes)}

\item{ontology_type}{A character string indicating the ontology type to be
used. It can be one of the following: "KEGG" (default), "GO", "MKEGG".}

\item{organism}{A character string indicating the organism to be used. It can
be either "human" (default) or "mouse".}
}
\value{
A data frame containing the enriched terms and associated statistics
}
\description{
This function performs ontology analysis using clusterProfiler. The user
provides a set of important features (e.g., genes), the ontology type to be
used (e.g., "KEGG", "GO", "MKEGG"), and the organism (e.g., "human" or "mouse").
}
\examples{
# Example with human KEGG pathways
data(FilterData)
# Example with mouse KEGG Module (MKEGG) terms
ontology_results <- Ontology_Analysis(
  Important_Features <- FilterData$Important_Features,
  ontology_type <- "MKEGG",
  organism <- "mouse"
)
}
