% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistical_Filter.R
\name{Statistical_Filter}
\alias{Statistical_Filter}
\title{Apply a statistical filter to gene expression data}
\usage{
Statistical_Filter(data, Pvalue_md, Labels, n_genes_to_keep)
}
\arguments{
\item{data}{A matrix or data.frame object representing gene expression data. Rows represent genes, and columns represent samples.}

\item{Pvalue_md}{A character string indicating the statistical method to use for filtering. Can be one of "Waldtest", "BPglm", "Wilcoxon-Test" or "LRT".}

\item{Labels}{A character or factor vector specifying the label for each cell in \code{data}.}

\item{n_genes_to_keep}{An integer indicating the number of genes to keep after filtering.}
}
\value{
A list with the following components:
\item{FilterData}{A matrix or data.frame object representing the filtered gene expression data.}
\item{Important_Features}{A character vector of gene names that pass the statistical filter.}
\item{Statistical_Analysis}{A data.frame object containing the p-values and FDRs for all genes, as calculated by the selected statistical method.}
}
\description{
This function performs a statistical filtering method on gene expression data. It filters the genes based on their statistical significance calculated by various methods.
}
\examples{

data(ExampleDataset)
Statistical_Filter(ExampleDataset[1:10,],Pvalue_md="Waldtest",Labels=Labels,n_genes_to_keep = 100)

}
